fixed prerender error
